CHG: sending has now timeout, issue #5
Introduced a timeout parameter in the sending plugin, and into
the ImgPacker.packAndSend(). Switched communication model to ZMQ.PAIR
to allow peer-to-peer bi-directional communication. Sender now sends
only a header and waits for response from receiver, than the communication
continues as usual/as it was before. ADD exchange of fake image metadata.